feat: PRO-2455 split functions in routes
