+++ backend/src/routes/metadata-routes.ts
